# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Lead {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  subject: string | null;
  /** @nullable */
  serviceInterested: string | null;
  message: string;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: string;
}

export interface LeadInput {
  /** @minLength 1 */
  fullName: string;
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  phone: string;
  companyName?: string;
  subject?: string;
  serviceInterested?: string;
  /** @minLength 1 */
  message: string;
}

export interface LeadUpdateInput {
  status?: string;
  notes?: string;
}

export interface Shipment {
  id: number;
  trackingNumber: string;
  status: string;
  origin: string;
  destination: string;
  carrierName: string;
  estimatedDelivery: string;
  lastUpdate: string;
}

export interface ShipmentInput {
  /** @minLength 1 */
  trackingNumber: string;
  /** @minLength 1 */
  status: string;
  /** @minLength 1 */
  origin: string;
  /** @minLength 1 */
  destination: string;
  /** @minLength 1 */
  carrierName: string;
  estimatedDelivery: string;
  /** @minLength 1 */
  lastUpdate: string;
}

export interface ShipmentUpdateInput {
  /** @minLength 1 */
  trackingNumber?: string;
  /** @minLength 1 */
  status?: string;
  /** @minLength 1 */
  origin?: string;
  /** @minLength 1 */
  destination?: string;
  /** @minLength 1 */
  carrierName?: string;
  estimatedDelivery?: string;
  /** @minLength 1 */
  lastUpdate?: string;
}

export interface AdminLoginInput {
  /** @minLength 1 */
  email: string;
  /** @minLength 1 */
  password: string;
}

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  roleName: string;
}