# lib/api-zod/src/generated/types/adminUser.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface AdminUser {
  id: number;
  email: string;
  name: string;
  roleName: string;
}