# lib/api-zod/src/generated/types/lead.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Lead {
  id: number;
  fullName: string;
  email: string;
  phone: string;
  /** @nullable */
  companyName: string | null;
  /** @nullable */
  subject: string | null;
  /** @nullable */
  serviceInterested: string | null;
  message: string;
  status: string;
  /** @nullable */
  notes: string | null;
  createdAt: Date;
}