# lib/api-zod/src/generated/types/shipmentUpdateInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ShipmentUpdateInput {
  /** @minLength 1 */
  trackingNumber?: string;
  /** @minLength 1 */
  status?: string;
  /** @minLength 1 */
  origin?: string;
  /** @minLength 1 */
  destination?: string;
  /** @minLength 1 */
  carrierName?: string;
  estimatedDelivery?: Date;
  /** @minLength 1 */
  lastUpdate?: string;
}